Github Actions Resource Workflow (#4)
Publishes Docker images to GHCR and binary tarballs to the pages
instance of this repo providing resources that Spack requires for
bootstrapping/basic functionality.
This currently includes:
| Package | Reason |
|-------------
|----------------------------------------------------------------------------
|
| mingw64-w64 | cross compiler Linux->Windows binaries, used for
compiling other resources |
| GnuPG | Required for signing binaries/buildcache entries |
| file | Required for file introspection, relocation, buildcaches |

Each resources is built via a docker image and uploaded to the GHCR
under the namespace `ghcr.io/spack/windows-resources/<pkg-name>`. Images
are tagged by the revision hash of the updating commit, with the most
recent receiving the `latest` tag.

Those resources are additionally extracted from the docker images, and
published to the pages instance for this repository, where they are
downloadable via in-browser links, or basic web requests.
Only the most recently updated resources are available on the pages
instance to minimize the space consumed by the resources.